refactor(gui): extract the partition-editor modal into a shared widget
Step 1 of making the disk-layout modal reusable from the Restore tab. Pure
refactor: no behaviour change, no new capability.

The modal turned out to be almost free to extract — it touched only five pieces
of InspectTab state and two of its methods, so it was already a self-contained
widget over PartitionEditor that happened to live inside inspect_tab.rs. It now
lives in gui/partition_editor_modal.rs and takes what it needs as parameters,
following the shared-widget pattern CONTRIBUTING points at for size_mode_row.

Applying stays with the caller, which owns the device handle and worker thread;
the widget returns Stay / Close / Apply the way physical_disk_export already
does. `build_and_validate_edits` disappears — the widget calls
`PartitionEditor::build_and_validate` directly.

inspect_tab drops 359 lines of modal plus 220 of layout-bar helpers for a
14-line wrapper, and loses a now-unused import and method (CONTRIBUTING rule 1:
zero warnings).

Inertness checked rather than assumed: the extracted body issues an identical
multiset of egui widget calls to the original at HEAD. Plus 2729 tests, clippy,
and the GUI, os-stub and vintage-1.73 builds.

Next: a BuildNew mode with a table-type picker, then per-row source images.
