+++ src/gui/inspect_tab.rs
@@ -23,7 +23,6 @@ use rusty_backup::model::export_runner::{
 use rusty_backup::model::partition_editor::PartitionEditor;
 use rusty_backup::model::status::{BlockCacheScan, CacheStatus, InspectStatus};
 use rusty_backup::partition::editor as pt_editor;
-use rusty_backup::partition::type_catalog;
 use rusty_backup::partition::{
     self, detect_alignment, PartitionAlignment, PartitionInfo, PartitionTable,
 };
@@ -1469,368 +1468,17 @@ impl InspectTab {
     }
 
     fn show_editor_popup(&mut self, ui: &mut egui::Ui, ctx: &mut TabContext) {
-        let mut open = true;
-        let mut apply_requested = false;
-
-        let kind = self
-            .partition_table
-            .as_ref()
-            .map(type_catalog::kind_of)
-            .unwrap_or(type_catalog::TableKind::Other);
-        let table_type = kind.label();
-
-        egui::Window::new("Edit Partition Table")
-            .open(&mut open)
-            .resizable(true)
-            .default_width(760.0)
-            .show(ui.ctx(), |ui| {
-                ui.label(format!("Table type: {}", table_type));
-                ui.add_space(4.0);
-
-                // Before / After disk layout. "Current" reads the loaded
-                // partition list, "After" walks the editor entries +
-                // add-partition inputs so the bar updates live as the user
-                // edits sizes, marks rows deleted, or fills in the
-                // Add Partition fields.
-                show_editor_disk_layout_bars(ui, &self.partitions, &self.editor, kind);
-                ui.add_space(8.0);
-
-                // Partition entry grid
-                egui::Grid::new("editor_grid")
-                    .striped(true)
-                    .min_col_width(50.0)
-                    .show(ui, |ui| {
-                        ui.label(egui::RichText::new("#").strong());
-                        ui.label(egui::RichText::new("Type").strong());
-                        ui.label(egui::RichText::new("Start LBA").strong());
-                        ui.label(egui::RichText::new("Size Mode").strong());
-                        ui.label(egui::RichText::new("Size (MiB)").strong());
-                        ui.label(egui::RichText::new("Boot").strong());
-                        ui.label(egui::RichText::new("").strong());
-                        ui.end_row();
-
-                        for i in 0..self.editor.entries.len() {
-                            // Copy values we need for display to avoid holding
-                            // an immutable borrow across mutable TextEdit borrows.
-                            let idx = self.editor.entries[i].index;
-                            let deleted = self.editor.entries[i].deleted;
-                            let is_ext = self.editor.entries[i].is_extended_container;
-                            let is_logical = self.editor.entries[i].is_logical;
-                            let start_lba = self.editor.entries[i].start_lba;
-                            let size_bytes = self.editor.entries[i].size_bytes;
-                            let bootable = self.editor.entries[i].bootable;
-                            let type_name = self.editor.entries[i].type_name.clone();
-
-                            if deleted {
-                                ui.label(
-                                    egui::RichText::new(format!("{}", idx))
-                                        .color(egui::Color32::GRAY)
-                                        .strikethrough(),
-                                );
-                                ui.label(
-                                    egui::RichText::new(&type_name)
-                                        .color(egui::Color32::GRAY)
-                                        .strikethrough(),
-                                );
-                                ui.label(
-                                    egui::RichText::new(format!("{}", start_lba))
-                                        .color(egui::Color32::GRAY),
-                                );
-                                ui.label("");
-                                ui.label(egui::RichText::new("deleted").color(egui::Color32::GRAY));
-                                ui.label("");
-                                if ui.small_button("Undo").clicked() {
-                                    self.editor.entries[i].deleted = false;
-                                }
-                                ui.end_row();
-                                continue;
-                            }
-
-                            if is_ext {
-                                ui.label(
-                                    egui::RichText::new(format!("{} (ext)", idx))
-                                        .color(egui::Color32::GRAY),
-                                );
-                                ui.label(
-                                    egui::RichText::new(&type_name).color(egui::Color32::GRAY),
-                                );
-                                ui.label(
-                                    egui::RichText::new(format!("{}", start_lba))
-                                        .color(egui::Color32::GRAY),
-                                );
-                                ui.label("");
-                                let size_mib = size_bytes as f64 / (1024.0 * 1024.0);
-                                ui.label(
-                                    egui::RichText::new(format!("{:.2}", size_mib))
-                                        .color(egui::Color32::GRAY),
-                                );
-                                ui.label("");
-                                ui.label("");
-                                ui.end_row();
-                                continue;
-                            }
-
-                            let label = if is_logical {
-                                format!("  {}", idx)
-                            } else {
-                                format!("{}", idx)
-                            };
-                            ui.label(label);
-
-                            // Type field: free-form text plus a catalog picker.
-                            type_field(
-                                ui,
-                                kind,
-                                &format!("ed_type_{}", i),
-                                &mut self.editor.entries[i].type_text,
-                            );
-
-                            // Start LBA (read-only)
-                            ui.label(format!("{}", start_lba));
-
-                            // Size-mode radios (Original / Minimum / Custom).
-                            // Selecting a non-Custom mode re-stamps
-                            // `size_text` to the canonical "{:.2}" MiB for
-                            // that target. Minimum is hidden when the
-                            // per-partition min size isn't known (the
-                            // editor doesn't perform its own FS analysis;
-                            // it relies on whatever min sizes were probed
-                            // by the inspect tab and seeded in via
-                            // `seed_from_with_minimums`).
-                            {
-                                use rusty_backup::model::size_mode::SizeMode;
-                                let minimum_size = self.editor.entries[i].minimum_size;
-                                let original_size = size_bytes;
-                                let prev = self.editor.entries[i].choice;
-                                ui.horizontal(|ui| {
-                                    ui.radio_value(
-                                        &mut self.editor.entries[i].choice,
-                                        SizeMode::Original,
-                                        "Original",
-                                    );
-                                    if minimum_size > 0 && minimum_size < original_size {
-                                        ui.radio_value(
-                                            &mut self.editor.entries[i].choice,
-                                            SizeMode::Minimum,
-                                            "Minimum",
-                                        );
-                                    }
-                                    ui.radio_value(
-                                        &mut self.editor.entries[i].choice,
-                                        SizeMode::Custom,
-                                        "Custom",
-                                    );
-                                });
-                                if self.editor.entries[i].choice != prev {
-                                    match self.editor.entries[i].choice {
-                                        SizeMode::Original => {
-                                            self.editor.entries[i].size_text = format!(
-                                                "{:.2}",
-                                                original_size as f64 / (1024.0 * 1024.0),
-                                            );
-                                        }
-                                        SizeMode::Minimum => {
-                                            self.editor.entries[i].size_text = format!(
-                                                "{:.2}",
-                                                minimum_size as f64 / (1024.0 * 1024.0),
-                                            );
-                                        }
-                                        _ => {}
-                                    }
-                                }
-                            }
-
-                            // Size (MiB) - editable when Custom; otherwise
-                            // displayed as a read-only field stamped by the
-                            // radio above.
-                            let size_editable = matches!(
-                                self.editor.entries[i].choice,
-                                rusty_backup::model::size_mode::SizeMode::Custom,
-                            );
-                            let size_id = format!("ed_size_{}", i);
-                            ui.add_enabled(
-                                size_editable,
-                                egui::TextEdit::singleline(&mut self.editor.entries[i].size_text)
-                                    .desired_width(80.0)
-                                    .id(egui::Id::new(&size_id)),
-                            );
-
-                            // Bootable cell — table-type-specific.
-                            // MBR carries a checkbox (legacy behaviour);
-                            // RDB renders explicit "boot" / "no boot" radio
-                            // buttons so the choice is impossible to flip
-                            // accidentally. GPT / APM have no per-partition
-                            // bootable bit, so the cell is read-only there.
-                            if table_type == "MBR" {
-                                ui.checkbox(&mut self.editor.entries[i].bootable, "");
-                            } else if table_type == "RDB" {
-                                ui.horizontal(|ui| {
-                                    let tip = "Whether this RDB partition is eligible to boot. \
-                                               Multiple partitions can be set to boot at once — \
-                                               the Amiga ROM picks the one with the highest boot \
-                                               priority among them.";
-                                    let mut val = self.editor.entries[i].bootable;
-                                    if ui
-                                        .radio_value(&mut val, true, "boot")
-                                        .on_hover_text(tip)
-                                        .changed()
-                                    {
-                                        self.editor.entries[i].bootable = val;
-                                    }
-                                    if ui
-                                        .radio_value(&mut val, false, "no boot")
-                                        .on_hover_text(tip)
-                                        .changed()
-                                    {
-                                        self.editor.entries[i].bootable = val;
-                                    }
-                                });
-                            } else {
-                                ui.label(if bootable { "Yes" } else { "" });
-                            }
-
-                            // Delete button
-                            if !is_logical {
-                                if ui
-                                    .small_button("Delete")
-                                    .on_hover_text("Mark partition for deletion")
-                                    .clicked()
-                                {
-                                    self.editor.entries[i].deleted = true;
-                                }
-                            } else {
-                                ui.label("");
-                            }
-
-                            ui.end_row();
-                        }
-                    });
-
-                ui.add_space(8.0);
-
-                // Add partition section. Open by default — reaching this
-                // popup at all usually means the user wants to allocate
-                // something, and a collapsed section hid the only way to do
-                // it.
-                egui::CollapsingHeader::new("Add Partition")
-                    .default_open(true)
-                    .show(ui, |ui| {
-                        // Free gaps, so the user can fill a hole left by a
-                        // deleted partition without hand-computing the LBA.
-                        let gaps = self.editor.free_gaps();
-                        if gaps.is_empty() {
-                            ui.label(
-                                egui::RichText::new("No unallocated space on this disk.").weak(),
-                            );
-                        } else {
-                            ui.horizontal_wrapped(|ui| {
-                                ui.label("Free space:");
-                                for gap in &gaps {
-                                    let label = format!(
-                                        "{} at LBA {}",
-                                        partition::format_size(gap.size_bytes),
-                                        gap.start_lba,
-                                    );
-                                    if ui
-                                        .button(label)
-                                        .on_hover_text("Fill this gap with the new partition")
-                                        .clicked()
-                                    {
-                                        self.editor.add_start_lba = gap.start_lba.to_string();
-                                        self.editor.add_size_mb =
-                                            format!("{}", (gap.size_bytes / (1024 * 1024)).max(1));
-                                    }
-                                }
-                            });
-                        }
-
-                        ui.horizontal(|ui| {
-                            ui.label("Start LBA:");
-                            ui.add(
-                                egui::TextEdit::singleline(&mut self.editor.add_start_lba)
-                                    .desired_width(90.0),
-                            )
-                            .on_hover_text(
-                                "First 512-byte sector of the new partition. It does not \
-                                 have to be past the existing partitions -- any free gap works.",
-                            );
-                            ui.label("Size (MiB):");
-                            ui.add(
-                                egui::TextEdit::singleline(&mut self.editor.add_size_mb)
-                                    .desired_width(80.0),
-                            );
-                            ui.label("Type:");
-                            type_field(ui, kind, "ed_add_type", &mut self.editor.add_type);
-                            if kind == type_catalog::TableKind::Mbr {
-                                ui.checkbox(&mut self.editor.add_bootable, "Bootable");
-                            }
-                            if ui.button("Add").clicked() {
-                                self.editor.add_entry_from_inputs(kind);
-                            }
-                        });
-
-                        ui.label(
-                            egui::RichText::new(format!(
-                                "Will be added as partition {} (slot assigned automatically; \
-                                 a slot freed by a Delete above is reused).",
-                                self.editor.next_entry_index(kind) + 1,
-                            ))
-                            .weak(),
-                        );
-                    });
-
-                ui.add_space(8.0);
-
-                // Show validation errors
-                for err in &self.editor.errors {
-                    ui.colored_label(egui::Color32::from_rgb(255, 100, 100), err);
-                }
-
-                // Show status
-                if let Some(status) = &self.editor.status {
-                    ui.colored_label(egui::Color32::from_rgb(100, 255, 100), status);
-                }
-
-                ui.add_space(4.0);
-
-                // Action buttons
-                ui.horizontal(|ui| {
-                    if ui.button("Validate").clicked() {
-                        self.build_and_validate_edits();
-                    }
-
-                    let can_apply = self.editor.errors.is_empty()
-                        && self.partition_table.is_some()
-                        && self.backup_folder_path.is_none();
-                    if ui
-                        .add_enabled(can_apply, egui::Button::new("Apply Changes"))
-                        .clicked()
-                    {
-                        // Build edits first
-                        self.build_and_validate_edits();
-                        if self.editor.errors.is_empty() && !self.editor.edits.is_empty() {
-                            apply_requested = true;
-                        }
-                    }
-
-                    if ui.button("Cancel").clicked() {
-                        self.editor_popup = false;
-                    }
-                });
-            });
-
-        if !open {
-            self.editor_popup = false;
-        }
-
-        if apply_requested {
-            self.apply_editor_changes(ctx);
-        }
-    }
-
-    fn build_and_validate_edits(&mut self) {
-        if let Some(table) = self.partition_table.as_ref() {
-            self.editor.build_and_validate(table);
+        let action = super::partition_editor_modal::show(
+            ui,
+            &mut self.editor,
+            &self.partitions,
+            self.partition_table.as_ref(),
+            self.backup_folder_path.is_none(),
+        );
+        match action {
+            super::partition_editor_modal::Action::Stay => {}
+            super::partition_editor_modal::Action::Close => self.editor_popup = false,
+            super::partition_editor_modal::Action::Apply => self.apply_editor_changes(ctx),
         }
     }
 
@@ -5945,226 +5593,6 @@ fn show_export_disk_layout_bars(
         ui.set_width(after_w.max(60.0));
         PartitionBar {
             segments: after,
-            show_inline_labels: true,
-            show_legend: true,
-        }
-        .show(ui);
-    });
-}
-
-/// Partition-type entry field: free-form text plus a catalog dropdown and the
-/// resolved type name.
-///
-/// The text field stays authoritative — the dropdown only stamps a value into
-/// it — so a type that isn't in the catalog is still reachable by typing it.
-fn type_field(ui: &mut egui::Ui, kind: type_catalog::TableKind, id: &str, text: &mut String) {
-    ui.horizontal(|ui| {
-        ui.add(
-            egui::TextEdit::singleline(text)
-                .desired_width(if kind == type_catalog::TableKind::Gpt {
-                    250.0
-                } else {
-                    90.0
-                })
-                .id(egui::Id::new(id)),
-        )
-        .on_hover_text(kind.field_hint());
-
-        let choices = type_catalog::choices(kind);
-        if !choices.is_empty() {
-            egui::ComboBox::from_id_salt(format!("{}_pick", id))
-                .selected_text("Pick...")
-                .width(230.0)
-                .show_ui(ui, |ui| {
-                    for choice in choices {
-                        let selected =
-                            type_catalog::describe(kind, text).is_some_and(|l| l == choice.label);
-                        if ui
-                            .selectable_label(
-                                selected,
-                                format!("{}  -  {}", choice.value, choice.label),
-                            )
-                            .clicked()
-                        {
-                            *text = choice.value.to_string();
-                        }
-                    }
-                })
-                .response
-                .on_hover_text("Choose a well-known partition type");
-        }
-
-        match type_catalog::describe(kind, text) {
-            Some(label) => {
-                ui.label(egui::RichText::new(label).weak());
-            }
-            None if !text.trim().is_empty() => {
-                ui.label(egui::RichText::new("custom").weak().italics());
-            }
-            None => {}
-        }
-    });
-}
-
-/// A partition's place on the disk, for the editor's layout bars.
-struct PlacedSegment {
-    label: String,
-    fs: String,
-    start_byte: u64,
-    size_bytes: u64,
-    color_index: usize,
-}
-
-/// Lay `placed` out across a `disk_size`-byte disk, inserting `Free` segments
-/// for every unallocated gap.
-///
-/// Segments are painted at their real byte offsets, so a partition added into
-/// a hole between two others renders in that hole rather than at the end of
-/// the bar. Input must be sorted by `start_byte`; an entry that starts before
-/// the previous one ends (an overlap the user hasn't validated away yet) is
-/// drawn straight after it instead of producing a negative gap.
-fn place_segments(placed: &[PlacedSegment], disk_size: u64) -> Vec<super::partition_bar::Segment> {
-    use super::partition_bar::{Segment, SegmentKind};
-
-    let mut segments = Vec::new();
-    let mut cursor = 0u64;
-    for p in placed {
-        if p.start_byte > cursor {
-            segments.push(Segment {
-                label: String::new(),
-                fs: String::new(),
-                size_bytes: p.start_byte - cursor,
-                kind: SegmentKind::Free,
-            });
-        }
-        segments.push(Segment {
-            label: p.label.clone(),
-            fs: p.fs.clone(),
-            size_bytes: p.size_bytes,
-            kind: SegmentKind::Partition {
-                color_index: p.color_index,
-            },
-        });
-        cursor = cursor.max(p.start_byte.saturating_add(p.size_bytes));
-    }
-    if disk_size > cursor {
-        segments.push(Segment {
-            label: String::new(),
-            fs: String::new(),
-            size_bytes: disk_size - cursor,
-            kind: SegmentKind::Free,
-        });
-    }
-    segments
-}
-
-/// Render the Current vs After PartitionBar pair inside the partition-table
-/// editor popup.
-///
-/// Both bars span the whole disk and place every partition at its real start
-/// LBA, with unallocated gaps drawn as free space — so the bar answers "where
-/// on the disk does this land", not merely "in what order are the entries".
-/// "Current" mirrors the loaded `partitions` list; "After" walks the editor's
-/// working entries plus the pending Add-Partition row.
-fn show_editor_disk_layout_bars(
-    ui: &mut egui::Ui,
-    partitions: &[PartitionInfo],
-    editor: &rusty_backup::model::partition_editor::PartitionEditor,
-    kind: type_catalog::TableKind,
-) {
-    use super::partition_bar::PartitionBar;
-
-    // Color by partition index, not by position, so a row keeps its color
-    // between the two bars even when an insertion shifts disk order.
-    let color_of = |index: usize| index;
-
-    let mut current: Vec<PlacedSegment> = partitions
-        .iter()
-        .filter(|p| !p.is_extended_container)
-        .map(|p| PlacedSegment {
-            label: format!("Partition {}", p.index + 1),
-            fs: p.type_name.clone(),
-            start_byte: p.byte_offset(),
-            size_bytes: p.size_bytes,
-            color_index: color_of(p.index),
-        })
-        .collect();
-    current.sort_by_key(|p| p.start_byte);
-
-    let mut after: Vec<PlacedSegment> = editor
-        .working_layout()
-        .into_iter()
-        .map(|p| PlacedSegment {
-            label: if p.is_new {
-                format!("New (part {})", p.index + 1)
-            } else {
-                format!("Partition {}", p.index + 1)
-            },
-            fs: p.type_name.clone(),
-            start_byte: p.start_byte(),
-            size_bytes: p.size_bytes,
-            color_index: color_of(p.index),
-        })
-        .collect();
-    // Pending Add-Partition inputs preview in place, before the user commits
-    // them with the Add button.
-    if let Some(pending) = editor.pending_add(kind) {
-        after.push(PlacedSegment {
-            label: "New".to_string(),
-            fs: pending.type_name.clone(),
-            start_byte: pending.start_byte(),
-            size_bytes: pending.size_bytes,
-            color_index: color_of(pending.index),
-        });
-    }
-    after.sort_by_key(|p| p.start_byte);
-
-    // Overlaps or entries running past the end make the working layout wider
-    // than the disk; scale both bars to that so the overflow stays visible
-    // instead of being silently clipped.
-    let allocated = |segs: &[PlacedSegment]| -> u64 { segs.iter().map(|p| p.size_bytes).sum() };
-    let extent = |segs: &[PlacedSegment]| -> u64 {
-        segs.iter()
-            .map(|p| p.start_byte.saturating_add(p.size_bytes))
-            .max()
-            .unwrap_or(0)
-    };
-    let disk_size = editor
-        .disk_size
-        .max(extent(&current))
-        .max(extent(&after))
-        .max(1);
-
-    let available_width = ui.available_width().max(120.0);
-    let current_free = disk_size.saturating_sub(allocated(&current));
-    let after_free = disk_size.saturating_sub(allocated(&after));
-
-    ui.label(format!(
-        "Current  (disk {}, {} allocated, {} free):",
-        partition::format_size(disk_size),
-        partition::format_size(allocated(&current)),
-        partition::format_size(current_free),
-    ));
-    ui.scope(|ui| {
-        ui.set_width(available_width);
-        PartitionBar {
-            segments: place_segments(&current, disk_size),
-            show_inline_labels: true,
-            show_legend: false,
-        }
-        .show(ui);
-    });
-
-    ui.add_space(4.0);
-    ui.label(format!(
-        "After  ({} allocated, {} free):",
-        partition::format_size(allocated(&after)),
-        partition::format_size(after_free),
-    ));
-    ui.scope(|ui| {
-        ui.set_width(available_width);
-        PartitionBar {
-            segments: place_segments(&after, disk_size),
             show_inline_labels: true,
             show_legend: true,
         }
